fix: WalletConnect renamed files, sign
