+++ apps/web/src/components/beacon/index.ts
@@ -1,4 +1,3 @@
 export * from "./BeaconProvider";
 export * from "./PermissionRequestModal";
-export * from "./SignPayloadRequestModal";
 export * from "./useHandleBeaconMessage";
